requirements.txt
streamlit==1.52.2
tensorflow==2.13.0
numpy==1.24.3
pillow==12.1.1
